+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v0.3.3 - 2026-09-04
+
+### What's Changed
+
+* chore: update CHANGELOG for v0.3.2 by @allnetru in https://github.com/allnetru/laravel-sharding/pull/47
+* fix: belongsTo on a shardable model lost the related type by @allnetru in https://github.com/allnetru/laravel-sharding/pull/48
+
+**Full Changelog**: https://github.com/allnetru/laravel-sharding/compare/v0.3.2...v0.3.3
+
 ## v0.3.2 - 2026-09-04
 
 ### What's Changed